separated to env.js
