+++ src/index.html
@@ -212,6 +212,7 @@
     <script src="./js/scripts.js" type="module"></script>
     <script src="./js/lighting.js" type="module"></script>
     <script src="./js/userInteractions.js" type="module"></script>
+    <script src="./js/env.js" type="module"></script>
     <script src="/node_modules/webm-writer/WebMWriter.js"></script>
 </body>
 </html>
